Fully removed Suites, Added methods to call for facilities values to be used for cart. Added CartItem and Shopping Cart. Basic functions of Rooms and Ballrooms Add to cart done.

+++ HotelBooking/JsonProperties/BallRoomResource.cs
@@ -11,6 +11,6 @@ namespace HotelBooking.JsonProperties
         public string ballroomType { get; set; }
         public List<BallroomId> ballroomId { get; set; }
         public int availability { get; set; }
-        public int cost { get; set; }
+        public double cost { get; set; }
     }
 }
