+++ HotelBooking/JsonProperties/HotelResource.cs
@@ -11,6 +11,6 @@ namespace HotelBooking.JsonProperties
         public string roomType { get; set; }
         public List<RoomId> roomId { get; set; }
         public int availability { get; set; }
-        public int cost { get; set; }
+        public double cost { get; set; }
     }
 }
